Added a central file for shared functions
