+++ fll_presenter/GUI_configurations/main_gui.py
@@ -9,6 +9,7 @@ import GUI_Elements.volume_control as vm
 import GUI_Elements.window_control as wc
 import GUI_Elements.timer as Timer
 
+import shared_functions
 
 system = platform.system()
 LASTCLICKMS = 0
@@ -32,12 +33,12 @@ def run_gui():
     timer.place(relx=.5, rely=.5,anchor="c")
 
     toggle_full_screen_button = ct.CTkButton(controller, text="Toggle\nFull\nScreen",
-    command=lambda:toggle_full_screen(timer_window),
+    command=lambda:shared_functions.toggle_full_screen(timer_window),
     width=110, height= 3, font=("Serif", 20))
     toggle_full_screen_button.pack(side="right", fill="y")
 
     update_color_button = ct.CTkButton(controller, text="Change\nTimer\nBackground",
-    command=lambda: timer_window.configure(background=str(open_color_menu(timer_window))),
+    command=lambda: timer_window.configure(background=str(shared_functions.open_color_menu(timer_window))),
     width=110, height= 3, font=("Serif", 20))
     update_color_button.pack(side="right", fill="y")
 
@@ -54,19 +55,19 @@ def run_gui():
     width=110, height=3, font=("Serif", 20))
 
     volume_up_button = ct.CTkButton(controller, text="Volume Up",
-    command=lambda: call_volume_control(True, controller, system),
+    command=lambda: shared_functions.call_volume_control(True, controller, system),
     width=110, height=3, font=("Serif", 20))
 
     volume_down_button = ct.CTkButton(controller, text="Volume Down",
-    command=lambda: call_volume_control(False, controller, system),
+    command=lambda: shared_functions.call_volume_control(False, controller, system),
     width=110, height=3, font=("Serif", 20))
 
     min_volume_slider = ct.CTkSlider(controller, width=20, from_=0, to=100,
-    number_of_steps=100, orientation="vertical", command=set_volume_minimum)
+    number_of_steps=100, orientation="vertical", command=shared_functions.set_volume_minimum)
     min_volume_slider.set(vm.get_volume_control()[0])
 
     max_volume_slider = ct.CTkSlider(controller, width=20, from_=0, to=100,
-    number_of_steps=100, orientation="vertical", command=set_volume_maximum)
+    number_of_steps=100, orientation="vertical", command=shared_functions.set_volume_maximum)
     max_volume_slider.set(vm.get_volume_control()[1])
 
 
@@ -85,10 +86,10 @@ def run_gui():
 
     if system == "Windows":
         application_selector = ct.CTkOptionMenu(controller, values=wc.get_window_executable_names(),
-        command=set_audio_application)
+        command=shared_functions.set_audio_application)
         application_selector.pack(side="left")
         application_selector.after(10000,
-        lambda:schedule_dropdown_update(controller, application_selector))
+        lambda:shared_functions.schedule_dropdown_update(controller, application_selector))
 
     timer_size = ct.CTkSlider(controller, width=20, from_=0, to=1000,
     orientation="vertical")
@@ -96,72 +97,10 @@ def run_gui():
     timer_size.set(200)
     timer_size.pack(side=("right"), fill='y')
 
-    timer_window.protocol("WM_DELETE_WINDOW", lambda:kill_windows(controller,timer_window))
-    controller.protocol("WM_DELETE_WINDOW", lambda:kill_windows(controller,timer_window))
+    timer_window.protocol("WM_DELETE_WINDOW", lambda:shared_functions.kill_windows(controller,timer_window))
+    controller.protocol("WM_DELETE_WINDOW", lambda:shared_functions.kill_windows(controller,timer_window))
 
     timer_window.mainloop()
 
 
-def schedule_dropdown_update(window, selector):
-    """Function to update the selector applications"""
-    selector.configure(values=wc.get_window_executable_names())
-    window.update()
-    selector.after(10000, lambda: schedule_dropdown_update(window, selector))
 
-
-def call_volume_control(option, controller, os):
-    """Function to call volume control with a 1 second cooldown"""
-    global LASTCLICKMS
-    if time.time() > (LASTCLICKMS + CLICKCOOLDOWN):
-        vm.volume_control(option, controller, os)
-        LASTCLICKMS = time.time()
-
-
-def set_volume_minimum(min_volume):
-    """Function to send the new volume minimum"""
-    vm.set_volume_control(min_volume, vm.get_volume_control()[1])
-
-def set_volume_maximum(max_volume):
-    """Function to send the new volume maximum"""
-    vm.set_volume_control(vm.get_volume_control()[0], max_volume)
-
-def set_audio_application(application):
-    """Function to send the new audio application"""
-    vm.set_volume_application(application)
-
-def kill_windows(controller, timer_window = None, kill_program = True):
-    """Function to kill the tkinter fll_presenter windows and exit the program"""
-    print("KILLING PROGRAM/WINDOW(S)")
-    controller.destroy()
-    if timer_window is not None:
-        timer_window.destroy()
-    if kill_program is True:
-        sys.exit()
-
-def open_color_menu(timer_window):
-    """Function to open a new color selector window for the color of the timer window"""
-    color = tkinter.colorchooser.askcolor(
-    title="Select a new color for the background of the display",
-                        initialcolor=timer_window['background'])[1]
-    if color is None:
-        return timer_window['background']
-    return color
-
-def toggle_full_screen(window):
-    """Function that uses a boolean to toggle full screen on and off for the argument window"""
-    global SCREENTOGGLE
-    if SCREENTOGGLE is False:
-        full_screen(window)
-        SCREENTOGGLE = True
-    elif SCREENTOGGLE is True:
-        exit_full_screen(window)
-        SCREENTOGGLE = False
-
-
-def full_screen(window):
-    """Function that fullscreens the supplied window"""
-    window.attributes("-fullscreen", True)
-
-def exit_full_screen(window):
-    """Function that exits fullscreen on the supplied window"""
-    window.attributes("-fullscreen", False)
